Moved resource stores to their own file
